Tried to modify 'Convert' button; discovered bug.
Tried to make it so that the 'Convert' button would change to 'Converting...' after being pressed but before the output was converted. However, while testing, I discovered that large inputs can cause an error that only shows up on the console log. It's likely because the parser is handling the text all at once, rather than breaking it up into chunks first.